Add Ranked & Superstars to netplay browser
The ranked button was moved to the "host" screen rather than being in the lobby itself. This might not be popular but will be needed for integrated matchmaking/ranked (also old system was a bit buggy anyway)

Superstar characters On/Off was also added to the host screen

Both of these were added to the netplay browser. Now people can see if a lobby is ranked and stars on/off before joining. hopefully this makes matchmaking easier before it's automated.

I also cleaned up the lobby too. Moved the coin flipper around and changed some color schemes